# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/user":{"post":{"tags":["user"],"summary":"Create User","operationId":"create_user_user_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDisplay"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/avatar/{id}":{"post":{"tags":["user"],"summary":"Upload Avatar","operationId":"upload_avatar_user_avatar__id__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_avatar_user_avatar__id__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/user/avatar":{"post":{"tags":["user"],"summary":"Update Avatar","operationId":"update_avatar_user_avatar_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvatarBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/post/all":{"get":{"tags":["post"],"summary":"Posts","operationId":"posts_post_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Posts Post All Get","type":"array","items":{"$ref":"#/components/schemas/PostDisplay"}}}}}}}},"/post/image":{"post":{"tags":["post"],"summary":"Upload File","operationId":"upload_file_post_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_post_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/post/delete/{id}":{"get":{"tags":["post"],"summary":"Delete","operationId":"delete_post_delete__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/login":{"post":{"tags":["authentication"],"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/all/{post_id}":{"get":{"tags":["comment"],"summary":"Comments","operationId":"comments_comment_all__post_id__get","parameters":[{"required":true,"schema":{"title":"Post Id","type":"integer"},"name":"post_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment":{"post":{"tags":["comment"],"summary":"Create","operationId":"create_comment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AvatarBase":{"title":"AvatarBase","required":["avatar_url","avatar_url_type"],"type":"object","properties":{"avatar_url":{"title":"Avatar Url","type":"string"},"avatar_url_type":{"title":"Avatar Url Type","type":"string"}}},"Body_login_login_post":{"title":"Body_login_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Body_upload_avatar_user_avatar__id__post":{"title":"Body_upload_avatar_user_avatar__id__post","required":["image"],"type":"object","properties":{"image":{"title":"Image","type":"string","format":"binary"}}},"Body_upload_file_post_image_post":{"title":"Body_upload_file_post_image_post","required":["image"],"type":"object","properties":{"image":{"title":"Image","type":"string","format":"binary"}}},"Comment":{"title":"Comment","required":["text","username","timestamp"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"username":{"title":"Username","type":"string"},"timestamp":{"title":"Timestamp","type":"string","format":"date-time"}}},"CommentBase":{"title":"CommentBase","required":["username","text","post_id"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"text":{"title":"Text","type":"string"},"post_id":{"title":"Post Id","type":"integer"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"PostDisplay":{"title":"PostDisplay","required":["id","image_url","image_url_type","caption","timestamp","user","comments"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"image_url":{"title":"Image Url","type":"string"},"image_url_type":{"title":"Image Url Type","type":"string"},"caption":{"title":"Caption","type":"string"},"timestamp":{"title":"Timestamp","type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/User"},"comments":{"title":"Comments","type":"array","items":{"$ref":"#/components/schemas/Comment"}}}},"User":{"title":"User","required":["username"],"type":"object","properties":{"username":{"title":"Username","type":"string"}}},"UserBase":{"title":"UserBase","required":["username","email","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"},"avatar_url":{"title":"Avatar Url","type":"string"},"avatar_url_type":{"title":"Avatar Url Type","type":"string"}}},"UserDisplay":{"title":"UserDisplay","required":["username","email"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"email":{"title":"Email","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}